# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Lavoro Ghana Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # --- USERS ---
...
  /freelancers:
    get:
      operationId: listFreelancers
      tags: [freelancers]
      summary: List freelancer profiles
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: skill
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: minRate
          in: query
          schema:
            type: number
        - name: maxRate
          in: query
          schema:
            type: number
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Freelancers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreelancerListResponse"
  /freelancers/me:
    get:
      operationId: getMyFreelancerProfile
      tags: [freelancers]
      summary: Get my freelancer profile
      responses:
        "200":
          description: My freelancer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreelancerProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: upsertFreelancerProfile
      tags: [freelancers]
      summary: Create or update my freelancer profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FreelancerProfileInput"
      responses:
        "200":
          description: Profile saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreelancerProfile"
  /freelancers/{userId}:
    get:
      operationId: getFreelancerProfile
      tags: [freelancers]
      summary: Get freelancer profile by user ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Freelancer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreelancerProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # --- CLIENT PROFILES ---
  /clients/me:
    get:
      operationId: getMyClientProfile
      tags: [clients]
      summary: Get my client profile
      responses:
        "200":
          description: My client profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: upsertClientProfile
      tags: [clients]
      summary: Create or update my client profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientProfileInput"
      responses:
        "200":
          description: Profile saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientProfile"
  /clients/{userId}:
    get:
      operationId: getClientProfile
      tags: [clients]
      summary: Get client profile by user ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # --- JOBS ---
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RegisterInput:
      type: object
      required: [name, email, password, phone, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [freelancer, client, seller, admin]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"

    # --- USERS ---
    User:
      type: object
      required: [id, name, email, phone, role, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [freelancer, client, seller, admin]
        status:
          type: string
          enum: [active, suspended]
        avatarUrl:
          type: ["string", "null"]
        isVerified:
          type: boolean
        averageRating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
...
    FreelancerProfile:
      type: object
      required: [id, userId, user, skills, location, hourlyRate, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        user:
          $ref: "#/components/schemas/User"
        bio:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        location:
          type: string
        portfolioItems:
          type: array
          items:
            $ref: "#/components/schemas/PortfolioItem"
        experience:
          type: ["string", "null"]
        hourlyRate:
          type: ["number", "null"]
        projectRate:
          type: ["number", "null"]
        completedJobs:
          type: integer
        verificationStatus:
          type: string
          enum: [unverified, pending, verified]
        trustLabel:
          type: ["string", "null"]
        createdAt:
          type: string
    FreelancerProfileInput:
      type: object
      properties:
        bio:
          type: string
        skills:
          type: array
          items:
            type: string
        location:
          type: string
        experience:
          type: string
        hourlyRate:
          type: number
        projectRate:
          type: number
        portfolioItems:
          type: array
          items:
            $ref: "#/components/schemas/PortfolioItemInput"
    FreelancerListResponse:
      type: object
      required: [freelancers, total]
      properties:
        freelancers:
          type: array
          items:
            $ref: "#/components/schemas/FreelancerProfile"
        total:
          type: integer
    PortfolioItem:
      type: object
      required: [id, title]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        link:
          type: ["string", "null"]
    PortfolioItemInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        link:
          type: string

    # --- CLIENT PROFILES ---
    ClientProfile:
      type: object
      required: [id, userId, user, location, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        user:
          $ref: "#/components/schemas/User"
        businessName:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        location:
          type: string
        logoUrl:
          type: ["string", "null"]
        postedJobsCount:
          type: integer
        createdAt:
          type: string
    ClientProfileInput:
      type: object
      properties:
        businessName:
          type: string
        description:
          type: string
        location:
          type: string
        logoUrl:
          type: string

    # --- JOBS ---
...