# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Lavoro Ghana Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';
import type { UserStatus } from './userStatus';

export interface User {
  id: number;
  name: string;
  email: string;
  phone: string;
  role: UserRole;
  status: UserStatus;
  /** @nullable */
  avatarUrl?: string | null;
  isVerified?: boolean;
  /** @nullable */
  averageRating?: number | null;
  reviewCount?: number;
  createdAt: string;
}